Move validateDirtyProps util
